Adding a new blog under Misc. (#22)
Added my blog entry that describes my initial journey constructing an Azure AKS cluster with integrated container registry, service principals - and a second project to deploy the apps, do a docker build + push.

The article talks a lot about how the "real code" aspect transforms the IaC at a fundamental level, and what aspects of the SDLC are affected.  It is written from the perspective of a developer.  

Thanks for the consideration.